requirements.txt
pandas==2.2.0
pygame_ce==2.3.2
pymongo==4.6.1
python_bcrypt==0.3.2
scikit_learn==1.4.0
